requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
anthropic==0.36.2
